MAINT: Split dynamic-space classes into core/dynamic.py; retire shims (Phase 8)
Final phase of the core refactoring (devnotes/CoreRefactorDesign.md).

- Move DynamicBase, DynamicSpaceImpl and ItemSpaceImpl verbatim from
  core/space.py to new core/dynamic.py. space.py re-imports them
  permanently (runtime pickles record the historical
  modelx.core.space paths, and UserSpaceImpl inherits DynamicBase);
  the DynamicSpace/ItemSpace interface classes stay in space.py,
  relocated above UserSpaceImpl so the import can precede it. The one
  mechanical change: DynamicSpaceImpl._init_allargs resolves
  UserSpaceImpl through the space module object at runtime.
- Retire the Phase 1-3 compat aliases past their grace period:
  the eight idstr node-string helpers, SharedSpaceOperations and
  ReferenceManager (incl. the refs.py alias). model.py keeps only the
  pickle-relevant SpaceGraph and SpaceManager aliases (spmgr is a
  slot of every pickled Impl; the graph lives inside it).
- Delete the unused modelx/managers sketch package together with its
  only consumer modelx/tests/managers, and the stale core/members/
  remnant.

Full suite: 1082 passed, 6 skipped.
