+++ devnotes/CoreRefactorDesign.md
@@ -4,6 +4,7 @@
 - Revision 2026-07-13: **D-7 reversed — all mixins stay** (no mixin→component conversions). Old Phase 2 (TraceManager component) replaced by the self-observation fix (D-12); old Phase 7 (BaseSpaceImpl decomposition) dropped; selective invalidation is now Phase 7, file split Phase 8.
 - Revision 2026-07-15: **serializer compatibility narrowed** (user decision). Serializers 2–3 are being deprecated for *loading* and 2–5 for *writing* (in a separate session, independent of this refactoring). The saved-model compatibility surface for this refactoring is therefore **loading serializer_4–7 only**. (Phase 0 had found the serializer_2/3 readers already broken on master — `UserSpaceImpl.new_cells` gone; deprecation supersedes fixing them.)
 - Revision 2026-07-17 (Phase 7 as implemented): five refinements to §5.8. (1) **Granularity is per itemspace**, not per parent: itemspace I is cleared iff `dirty_spaces ∩ closure(I) ≠ ∅` with `closure(I)` = I's recorded tree dynbases + their MRO bases + the parent's nearest static UserSpace. The `*_nukes_at_parent` characterization tests flipped accordingly. (2) **Two dirty buckets**: `ChangeSet.dirty_spaces` (namespace-binding changes; intersected with the full closure) and `ChangeSet.dirty_bases` (formula changes/renames via `txn.mark_dirty_base`; intersected with the recorded dynbases only) — preserving the per-dynbase selectivity that `clear_subs_rootitems` gave formula/rename edits, including for the edited space's own itemspaces built on foreign dynbases. (3) The new `ItemSpaceImpl` slot (`tree_dynbases`) records **impl references, not idstr strings** — idstrs and MRO are resolved at invalidation time, so the closure follows renames of base spaces instead of going stale. (4) `DelSpace` now records the removed spaces in `dirty_spaces` (via `add_removed_space`), fixing an under-clearing hole where itemspaces built on a deleted space survived with a dangling dynbase. (5) **Cross-model invalidation** goes through the dirty spaces' `_dynamic_subs` back-pointers (idstrs are model-relative, so closure intersection is only valid within one model); root itemspaces of other models built on dirty spaces are cleared per root. `ChangeSet.cleared_subs`/`clear_subs_rootitems`/`DynamicBase.on_notify`/`on_namespace_change` are removed (subsumed/dead). Known pre-existing hole left for later: deleting a space does not delete the deleted space's *own* live itemspaces (their trace nodes linger); unchanged from before Phase 7.
+- Revision 2026-07-17 (Phase 8 as implemented): `DynamicBase`/`DynamicSpaceImpl`/`ItemSpaceImpl` moved to new `core/dynamic.py`; the `DynamicSpace`/`ItemSpace` *interface* classes stay in `space.py` but are relocated above `UserSpaceImpl`, and `space.py` imports the moved impls mid-module right before `UserSpaceImpl` (which inherits `DynamicBase`) — that import is the permanent pickle alias (`modelx.core.space.ItemSpaceImpl` etc. resolve forever). One mechanical change in the moved code: `DynamicSpaceImpl._init_allargs` looks up `UserSpaceImpl` through the space module object at runtime. Alias retirement: `model.py` keeps `SpaceGraph` and `SpaceManager` only (pickle-relevant — `spmgr` is a slot of every pickled Impl, the graph lives inside it); the eight idstr node-helper aliases, `SharedSpaceOperations`, and `ReferenceManager` (incl. the `refs.py` alias) are deleted. `modelx/managers` deleted together with its only consumer `modelx/tests/managers`; `core/members/` (untracked pycache remnant) deleted; `core/inheritance/__pycache__` contained only current-module caches by this point — nothing stale left to remove.
 - Purpose: self-contained guide for executing the refactoring one phase per session. A session picking up any phase should need only this document plus the source tree.
 - Line numbers below refer to the tree as of the commit above; re-verify before editing, but the structural claims were all confirmed against source when this was written.
